+++ modules/library/main/src/main/java/org/geotools/renderer/crs/WorldVanDerGrintenIHandlerFactory.java
@@ -47,8 +47,8 @@ public class WorldVanDerGrintenIHandlerFactory implements ProjectionHandlerFacto
         if (mapProjection instanceof WorldVanDerGrintenI) {
             ReferencedEnvelope validArea =
                     new ReferencedEnvelope(
-                            -Double.MAX_VALUE,
-                            Double.MAX_VALUE,
+                            -Integer.MAX_VALUE,
+                            Integer.MAX_VALUE,
                             -90,
                             90,
                             DefaultGeographicCRS.WGS84);
